+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 - Added notification system
 - Added ClientOverlay superclass
 - Added change-handlers to setting delegates
+- Added color setting
+- Added Block ESP module
+- Added list setting
+- Added `build` workflow
 
 # 0.1.0
 - Initial Commit